Add Data handling to `ConnectorChangeEvent` and refactor `ActionEvent` and `ConnectorChangeEvent` to use a common base class. This can be convenient for sharing data processing functions within a plugin, since they are so similar, eg. by using the base class for method arguments.
